# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CMM API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "CMM env"
        }
    ],
    "paths": {
...
        "/beneficiaries": {
            "get": {
                "tags": [
                    "Beneficiary"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "getBeneficiary",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "paginate",
                        "required": false,
                        "schema": {
                            "title": "Paginate",
                            "description": "The Paginate data",
                            "type": "boolean",
                            "example": "true"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string",
                            "example": "id"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string",
                            "example": "desc"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Sort order field",
                        "schema": {
                            "title": "perPage",
                            "description": "The unique identifier of a curso resource",
                            "type": "number",
                            "default": "10"
                        }
                    },
                    {
                        "name": "dataSearch",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "Search data",
                            "type": "string"
                        }
                    },
                    {
                        "name": "dataFilter",
                        "in": "query",
                        "description": "turno resource name",
                        "required": false,
                        "schema": {
                            "description": "The unique identifier of a turno resource",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Beneficiarys all."
                    },
                    "default": {
                        "description": "error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Beneficiary"
                ],
                "summary": "Creates a new Beneficiary",
                "description": "Creates a new Beneficiary",
                "operationId": "5212eb3663d5e1d59ba0dd27fb6df6d3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Beneficiary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The Beneficiary resource created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "an \"unexpected\" error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/beneficiaries/{id}": {
            "get": {
                "tags": [
                    "Beneficiary"
                ],
                "summary": "Get Beneficiary information",
                "description": "Returns Beneficiary data",
                "operationId": "getBeneficiaryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Beneficiary id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Beneficiary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Beneficiary"
                ],
                "summary": "Update existing Beneficiary",
                "description": "Returns updated Beneficiary data",
                "operationId": "updateBeneficiary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Beneficiary id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Beneficiary"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Beneficiary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Beneficiary"
                ],
                "summary": "Delete existing Beneficiary",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteBeneficiary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Beneficiary id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Beneficiary": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "properties": {
                            "name": {
                                "description": "The Beneficiary name",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "last_name": {
                                "description": "The Beneficiary last_name",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "document_number": {
                                "description": "The Beneficiary document_number",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "email": {
                                "description": "The Beneficiary email",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "phone_number": {
                                "description": "The Beneficiary phone_number",
                                "required": [
                                    "true"
                                ],
                                "type": "string"
                            },
                            "user_created_id": {
                                "description": "The Users crete",
                                "required": [
                                    "true"
                                ],
                                "type": "number",
                                "example": 1
                            },
                            "user_updated_id": {
                                "description": "The Users update",
                                "required": [
                                    "true"
                                ],
                                "type": "number",
                                "example": 1
                            }
                        }
                    }
                ]
            },
...
            "User": {
                "properties": {
                    "name": {
                        "description": "The user name",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "last_name": {
                        "description": "The Users password",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "email": {
                        "description": "The Users email",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "roles": {
                        "description": "The Product attribute_types",
                        "required": [
                            "false"
                        ],
                        "type": "array",
                        "items": {
                            "properties": {
                                "role_id": {
                                    "type": "number"
                                },
                                "branch_office_id": {
                                    "type": "number"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "password": {
                        "description": "The Users password",
                        "required": [
                            "true"
                        ],
                        "type": "string"
                    },
                    "user_created_id": {
                        "description": "The Users crete",
                        "required": [
                            "true"
                        ],
                        "type": "number",
                        "example": 1
                    },
                    "user_updated_id": {
                        "description": "The Users update",
                        "required": [
                            "true"
                        ],
                        "type": "number",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}